+++ max_div/benchmark_problems.py
@@ -1,0 +1,14 @@
+from ._core.benchmark_problems import BenchmarkProblem, BenchmarkProblemFactory
+
+__all__ = [
+    "BenchmarkProblem",
+    "BenchmarkProblemFactory",
+]
+
+
+# --- module patching ---------------------------------------------------------
+# This ensures that the user sees all re-exported names as belonging to this module, rather than their original
+from max_div._core._utils.api import patch_modules  # noqa: E402
+
+patch_modules(globals(), __all__, __name__)
+del patch_modules