+++ max_div/problem.py
@@ -1,0 +1,15 @@
+from ._core.constraints import Constraint
+from ._core.problem import MaxDivProblem
+
+__all__ = [
+    "Constraint",
+    "MaxDivProblem",
+]
+
+
+# --- module patching ---------------------------------------------------------
+# This ensures that the user sees all re-exported names as belonging to this module, rather than their original
+from max_div._core._utils.api import patch_modules  # noqa: E402
+
+patch_modules(globals(), __all__, __name__)
+del patch_modules